requirements.txt
Flask
qrcode
segno
Pillow
pandas
xlsxwriter
